Class Portfolio
This class uses a LinkedList to save all Investment. When anyone call the method evaluate, utilice one Iterator for evaluate all investment in the LinkedList.
For each bucle add the result to a var Money.
When is finished return this Money with the currency spicified in the parameters

Test for Portfolio:
-Params of addInvestment cannot be null.
-Method evaluate wich a Ticket not in StockExchange.
-Method evaluate with one investment.
-Method evaluate with two investment.